Add files via upload
This project contains data of around 32,000 people. It contains their ages, salaries, years of experience, credit history, etc. The goal of the project was to run a Logistic Regression model to find whether or not a person is eligible for the loan they are applying for or not. It was found that the model could predict whether a person was eligible accurately 79% of the time.